+++ azimuth/utils/project.py
@@ -34,6 +34,9 @@ def load_dataset_from_config(azimuth_config: AzimuthConfig) -> DatasetDict:
         ValueError if neither validation or test exists in the dataset
 
     """
+    if azimuth_config.dataset is None:
+        raise ValueError("No dataset configured.")
+
     dataset_in_config: DatasetDict = load_custom_object(
         azimuth_config.dataset, azimuth_config=azimuth_config
     )
@@ -56,7 +59,7 @@ def load_dataset_from_config(azimuth_config: AzimuthConfig) -> DatasetDict:
 def save_config(azimuth_config: AzimuthConfig):
     """Append config to config_history.jsonl to retrieve past configs."""
     # TODO https://stackoverflow.com/questions/2333872/how-to-make-file-creation-an-atomic-operation
-    with jsonlines.open(f"{azimuth_config.artifact_path}/config_history.jsonl", mode="a") as f:
+    with jsonlines.open(azimuth_config.get_config_history_path(), mode="a") as f:
         f.write(azimuth_config.dict())