+++ tests/test_routers/test_utterances.py
@@ -4,7 +4,11 @@
 from typing import List
 
 from fastapi import FastAPI
-from starlette.status import HTTP_200_OK, HTTP_404_NOT_FOUND, HTTP_422_UNPROCESSABLE_ENTITY
+from starlette.status import (
+    HTTP_200_OK,
+    HTTP_404_NOT_FOUND,
+    HTTP_422_UNPROCESSABLE_ENTITY,
+)
 from starlette.testclient import TestClient
 
 UTTERANCE_COUNT = 42